// src/app/models/startStudentRegModel.ts
export interface StudentStartRegModel {
  name: string;
  email: string;
  tc: string;
  no: string;
  location: string;
  collectionAmount: string;
  advancePay: string;
  educationHistory: Date;
  regHistory: Date;
  staff: string;
  dataSource: string;
  careerCounselor: string;
  interestedEducation: string;
  numberBills: string;
}